Extract interaction
