+++ frontend/scripts/react-components/tool/map/map-constants.js
@@ -1,0 +1,2 @@
+export const INDONESIA_MILL_LAYER_ID = 'indonesia_mill';
+export const getGeoIdName = layerId => (layerId === INDONESIA_MILL_LAYER_ID ? 'mill_name' : 'geoid');
